Conflicts:
	crossroad/src/main/java/fr/unice/polytech/al/trafficlight/crossroad/CrossroadModuleRunning.java
	utils/src/main/java/fr/unice/polytech/al/trafficlight/utils/RuleGroup.java
	utils/src/main/java/fr/unice/polytech/al/trafficlight/utils/Scenario.java

+++ utils/src/main/java/fr/unice/polytech/al/trafficlight/utils/RuleGroup.java
@@ -7,13 +7,14 @@ import java.util.Set;
  * Created by nathael on 26/10/16.
  */
 public class RuleGroup {
+
     private String id;
-    private long greenTime; // Green time in ms
+    private int greenTime; // Green time in s
     private Set<TrafficLightId> trafficLightList;
 
     public RuleGroup(){}
 
-    public RuleGroup(String id, long greenTime) {
+    public RuleGroup(String id, int greenTime) {
         this.id = id;
         this.greenTime = greenTime;
         trafficLightList = new HashSet<>();
@@ -46,9 +47,9 @@ public class RuleGroup {
 
     /**
      * Get how long is this group traffic lights's green time
-     * @return ms time
+     * @return s time
      */
-    public long getGreenTime() {
+    public int getGreenTime() {
         return greenTime;
     }
 
@@ -65,9 +66,9 @@ public class RuleGroup {
 
     /**
      * Change this group traffic lights's green time
-     * @param greenTime Green time in ms
+     * @param greenTime Green time in s
      */
-    public void setGreenTime(long greenTime) {
+    public void setGreenTime(int greenTime) {
         this.greenTime = greenTime;
     }
